+++ CHANGELOG.md
@@ -17,7 +17,8 @@ This is used for recording Invader's changes. This changelog is based on
         - Minimum error set to 0.0° in the first trigger
     - `weapons\plasma rifle\plasma rifle.weapon`
         - Error angle set to 0.5° - 5.0° in the first trigger
-- invader-extract: Globals tags are no longer extracted
+- invader-extract: Globals tags in non-multiplayer maps are no longer extracted
+  unless you use `-s`, and this is due to those tags having data stripped out
 
 ### Deprecated
 - invader-extract: Hidden values are no longer extracted by default. A compiler