+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 This is used for recording Invader's changes. This changelog is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.19.2] - 2019-11-24
 ### Added
 - invader-extract: Added recursive extraction